% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{live_past_events}
\alias{live_past_events}
\alias{live_upcoming_events}
\alias{live_draft_events}
\alias{live_private_events}
\alias{live_event}
\title{Get events}
\usage{
live_past_events(account.id = NULL, key = NULL,
  quiet = !interactive())

live_upcoming_events(account.id = NULL, key = NULL,
  quiet = !interactive())

live_draft_events(account.id = NULL, key = NULL,
  quiet = !interactive())

live_private_events(account.id = NULL, key = NULL,
  quiet = !interactive())

live_event(event.id = NULL, account.id = NULL, key = NULL)
}
\arguments{
\item{account.id}{your account id, can be setup via \code{\link{live_setup}}.}

\item{key}{your API key, optional, if passed overrides \code{\link{live_setup}}.}

\item{quiet}{Whether to print information about calls in the console.}

\item{event.id}{optional, if passed here overrides \code{\link{live_setup}}.}
}
\description{
Get information on events.
}
\examples{
\dontrun{
# setup account id
# to remove need to pass in subsequent calls
live_setup(
  key = "Xxx0xX0X0x0X0x", # your API token
  account.id = "xxXXxxXXxx"
)

past_events <- live_past_events()
event <- live_event(sample(past_events$id, 1))
}

}
