% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.R
\name{live_feed}
\alias{live_feed}
\alias{live_status}
\alias{live_image}
\alias{live_pending}
\title{Get feed items}
\usage{
live_feed(event.id = NULL, newer = 0, older = 25, offset.id = NULL,
  offset.type = NULL, account.id = NULL, key = NULL)

live_status(id, event.id = NULL, account.id = NULL, key = NULL)

live_image(id, event.id = NULL, account.id = NULL, key = NULL)

live_pending(event.id = NULL, account.id = NULL, key = NULL)
}
\arguments{
\item{event.id}{optional, if passed here overrides \code{\link{live_setup}}.}

\item{older, newer}{the number of posts to return older or newer than \code{offset}.
The maximum value of these parameters is limited to \code{25}.}

\item{offset.id}{the ID of the post returned in the previous call relative to which older or newer posts will be returned.
If no value is passed, this endpoint will fetch older posts based on \code{older} parameter value.}

\item{offset.type}{the type of post returned in the previous call relative to which older or newer posts will be returned.
Valid values are \code{status}, \code{image} and \code{video}.
This parameter is required if \code{offset.id} paramter is also present.}

\item{account.id}{optional, if passed here overrides \code{\link{live_setup}}.}

\item{key}{your API key, optional, if passed overrides \code{\link{live_setup}}.}

\item{id}{id of object to retrieve.}
}
\description{
Get feed items.
}
\examples{
\dontrun{
# setup account id
# to remove need to pass in subsequent calls
live_setup(
  key = "Xxx0xX0X0x0X0x", # your API token
  account.id = 1234567
  event.id = 1234567
)

feed <- live_feed()
}

}
