% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{live_account}
\alias{live_account}
\title{Get Account information}
\usage{
live_account(key = NULL)
}
\arguments{
\item{key}{your API key, optional, if passed overrides \code{\link{live_setup}}.}
}
\description{
Collect information aboutr your account.
}
\examples{
\dontrun{
live_setup(key = "Xxx0xX0X0x0X0x") # setup key

account <- live_account()
}

}
