% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3u8.R
\name{live_download}
\alias{live_download}
\title{Get videos}
\usage{
live_download(video.id, event.id = NULL, account.id = NULL,
  key = NULL)
}
\arguments{
\item{video.id}{id of video to download.}

\item{event.id}{optional, if passed here overrides \code{\link{live_setup}}.}

\item{account.id}{optional, if passed here overrides \code{\link{live_setup}}.}

\item{key}{your API key, optional, if passed overrides \code{\link{live_setup}}.}
}
\value{
a \code{list}, see \href{official documentation}{https://livestream.com/developers/docs/api/#download-video-urls}.
#' @export
}
\description{
Get videos of particular event.
}
\examples{
\dontrun{
# setup account id
# to remove need to pass in subsequent calls
live_setup(
  key = "Xxx0xX0X0x0X0x", # your API token
  account.id = 1234567
)

past_events <- live_past_events()

# random event
id <- past_events[sample(nrow(past_events), 1), "id"]
videos <- live_videos(id)

# get download url of random video
download_urls <- live_download(sample(videos$vods$data$data.id, 1))
}

}
