% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{live_events_gadget}
\alias{live_events_gadget}
\title{Events gadgets}
\usage{
live_events_gadget(account.id = NULL, key = NULL, which = c("past",
  "upcoming", "draft", "private"))
}
\arguments{
\item{account.id}{your account id, can be setup via \code{\link{live_setup}}.}

\item{key}{your API key, optional, if passed overrides \code{\link{live_setup}}.}

\item{which}{which events function to run.}
}
\description{
Events gadgets
}
\examples{
\dontrun{
# setup account id
# to remove need to pass in subsequent calls
live_setup(
  key = "Xxx0xX0X0x0X0x", # your API token
  account.id = 1909571 # Forum id
)

live_events_gadget()
}

}
