% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{live_search}
\alias{live_search}
\title{Search livestream}
\usage{
live_search(q, from = NULL, to = NULL, tags = NULL, account.id = NULL,
  key = NULL, quiet = !interactive())
}
\arguments{
\item{q}{query string.}

\item{from, to}{date range, optional.}

\item{tags}{a vector of event tag strings.}

\item{account.id}{defaults to all linked accounts unless an owned or linked account ID is specified.}

\item{key}{your API key, optional, if passed overrides \code{\link{live_setup}}.}

\item{quiet}{Whether to print information about calls in the console.}
}
\value{
a list of public and private events for all associated accounts (owned and linked) matched on query,
filters, and/or tags unless requested with a particular account ID. The response returns events in descending score
order based on our \href{ranking logic}{https://livestreamapis.com/apidocs/v3/#ranking}.
Tags can be added for each Event inside Livestream.com. We recommend using tags to easily search based on your
own custom categories or identifiers.
}
\description{
Search live stream
}
\examples{
\dontrun{
live_setup(key = "Xxx0xX0X0x0X0x") # setup key

results <- live_search(q = "Davos")
}

}
