% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{live_setup}
\alias{live_setup}
\alias{live_options}
\title{Setup}
\usage{
live_setup(key = NULL, account.id = NULL, event.id = NULL)

live_options()
}
\arguments{
\item{key}{your API key.}

\item{account.id}{your account id.}

\item{event.id}{event id.}
}
\description{
Setup your session.
}
\examples{
\dontrun{
live_setup(key = "Xxx0xX0X0x0X0x") # setup

live_options() # retrieve
}

}
